+++ paper/paper.bib
@@ -1,99 +1,109 @@
 @article{karakaya2024improved,
-  title={An Improved Cellular Genetic Algorithm with Machine-Coded Operators for Real-Valued Optimisation Problems},
-  author={Karakaya, Sevgi Akten and Satman, Mehmet Hakan},
-  journal={Journal of Engineering Research and Applied Science},
-  volume={13},
-  number={1},
-  pages={2500--2514},
-  year={2024}
+  title   = {An Improved Cellular Genetic Algorithm with Machine-Coded Operators for Real-Valued Optimisation Problems},
+  author  = {Karakaya, Sevgi Akten and Satman, Mehmet Hakan},
+  journal = {Journal of Engineering Research and Applied Science},
+  volume  = {13},
+  number  = {1},
+  pages   = {2500--2514},
+  year    = {2024}
+  % No DOI
 }
 
 @article{satman2013machine,
-  title={Machine coded genetic algorithms for real parameter optimization problems},
-  author={Satman, Mehmet Hakan},
-  journal={Gazi University Journal of Science},
-  volume={26},
-  number={1},
-  pages={85--95},
-  year={2013},
-  publisher={Gazi University}
+  title     = {Machine coded genetic algorithms for real parameter optimization problems},
+  author    = {Satman, Mehmet Hakan},
+  journal   = {Gazi University Journal of Science},
+  volume    = {26},
+  number    = {1},
+  pages     = {85--95},
+  year      = {2013},
+  publisher = {Gazi University}
+  % No DOI
 }
 
 @book{holland1975adaptation,
-  title={Adaptation in Natural and Artificial Systems},
-  author={Holland, John H.},
-  year={1975},
-  publisher={University of Michigan Press},
-  address={Ann Arbor, MI},
-  isbn={9780262581110}
+  title     = {Adaptation in Natural and Artificial Systems},
+  author    = {Holland, John H.},
+  year      = {1975},
+  publisher = {University of Michigan Press},
+  address   = {Ann Arbor, MI},
+  isbn      = {9780262581110}
+  % No DOI
 }
 
 @book{goldberg1989genetic,
-  title={Genetic Algorithms in Search, Optimization, and Machine Learning},
-  author={Goldberg, David E.},
-  year={1989},
-  publisher={Addison-Wesley},
-  address={Reading, MA},
-  isbn={9780201157673}
+  title     = {Genetic Algorithms in Search, Optimization, and Machine Learning},
+  author    = {Goldberg, David E.},
+  year      = {1989},
+  publisher = {Addison-Wesley},
+  address   = {Reading, MA},
+  isbn      = {9780201157673},
+  doi       = {10.5860/choice.27-0936}
 }
 
 @article{goldberg1991comparative,
-  title={A Comparative Analysis of Selection Schemes Used in Genetic Algorithms},
-  author={Goldberg, David E. and Deb, Kalyanmoy},
-  journal={Foundations of Genetic Algorithms},
-  volume={1},
-  pages={69--93},
-  year={1991},
-  publisher={Elsevier}
+  title     = {A Comparative Analysis of Selection Schemes Used in Genetic Algorithms},
+  author    = {Goldberg, David E. and Deb, Kalyanmoy},
+  journal   = {Foundations of Genetic Algorithms},
+  volume    = {1},
+  pages     = {69--93},
+  year      = {1991},
+  publisher = {Elsevier},
+  doi       = {10.1016/b978-0-08-050684-5.50008-2}
 }
 
 @inproceedings{manderick1989genetic,
-  title={The Genetic Algorithm and the Structure of the Fitness Landscape},
[…]
