+++ paper/paper.bib
@@ -76,7 +76,7 @@
 @book{eiben2003introduction,
   title     = {Introduction to Evolutionary Computing},
   author    = {Eiben, Agoston E. and Smith, James E.},
-  year      = {2003},
+  year      = {2015},
   publisher = {Springer},
   address   = {Berlin, Heidelberg},
   isbn      = {9783540401841},
@@ -102,8 +102,8 @@
   year      = {2008},
   publisher = {Springer},
   address   = {US},
-  isbn      = {978-0-387-77609-5}
-  % No DOI
+  isbn      = {978-0-387-77609-5},
+  doi       = {10.1007/978-0-387-77610-1_1}
 }
 
 @article{satman2019alpha_male_ga,
