+++ paper/paper.bib
@@ -6,7 +6,6 @@
   number  = {1},
   pages   = {2500--2514},
   year    = {2024}
-  % No DOI
 }
 
 @article{satman2013machine,
@@ -18,7 +17,6 @@
   pages     = {85--95},
   year      = {2013},
   publisher = {Gazi University}
-  % No DOI
 }
 
 @book{holland1975adaptation,
@@ -28,7 +26,6 @@
   publisher = {University of Michigan Press},
   address   = {Ann Arbor, MI},
   isbn      = {9780262581110}
-  % No DOI
 }
 
 @book{goldberg1989genetic,
@@ -59,7 +56,6 @@
   pages       = {143--150},
   year        = {1989},
   organization = {Morgan Kaufmann Publishers Inc.}
-  % No DOI
 }
 
 @book{michalewicz1996genetic,
@@ -92,7 +88,6 @@
   year      = {1993},
   publisher = {Morgan Kaufmann},
   address   = {California, CA, USA}
-  % No DOI
 }
 
 @book{alba2008cellular,
