### 2024-08-25 #7105 @SevgiAkten
Hi @JBorrow, thank you for accepting to be an editor for our submission.

The editorial bot suggests some potential reviewers by the similarity measure so we can say

- jmejia8
- ianhbell
- pySRURGS
- darioizzo
- jbuisine

may be invited as reviewers to our submission.

I also investigated the list of people and

- bhttchr6
- mpvanderschelling
- dc-luo
- vissarion
- devnums
- larkz
- jonbmartin
- iammix
- CeciliaCoelho
- Karangupta1994
- AlessandroPierro
- jbussemaker
- atismer
- ewu63
- mrava87
- parikshitbajpai

are suitable reviewers which have expertise on both Python and optimization.

Thank you in advance.
