### 2024-08-26 #7105 @JBorrow
Hi @jmejia8, I noticed you were the author of a similar package submitted to JOSS. Would you be interested in reviewing this one for us? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
