### 2024-08-26 #7105 @jonbmartin
Hi Josh,

I apologize, I am going to be quite busy for the next ~2 weeks, so I would
prefer not to commit to reviewing now. If you are still looking for a
reviewer at that time or need another round of reviews, I'd be happy to
help then!

Jonathan

On Mon, Aug 26, 2024 at 3:58 PM Josh Borrow ***@***.***>
wrote:

> Hi @jonbmartin <https://github.com/jonbmartin>, I found your name in the
> JOSS reviewer's database. Would you be available to review this submission?
> Our reviewer's guide is available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
