### 2024-09-03 #7105 @mrava87
@JBorrow thank you for the invite but unfortunately the topic of the paper is a bit too far from my expertise. I will need to pass on this 😞
