### 2024-09-09 #7105 @JBorrow
Hi @jbussemaker, I found your name in the JOSS reviewer's database. Would you be available to review this submission? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
