### 2024-10-07 #7322 @editorialbot
[REVIEW]: pycellga: A Python package for improved cellular genetic algorithms
**Submitting author:** @sevgiakten (<a href="http://orcid.org/0000-0001-9346-5795">Sevgi AKTEN KARAKAYA</a>)
**Repository:** https://github.com/SevgiAkten/pycellga
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.0
