### 2024-10-17 #7322 @JBorrow
Hi, a couple of reference notes:

- 9780262581110 is published by MIT Press, not Michigan (https://mitpress.mit.edu/9780262581110/adaptation-in-natural-and-artificial-systems/)
- [Proc. 4th Int. Conf. on Genetic Algorithms, 1991](https://cir.nii.ac.jp/all?q=Proc.%204th%20Int.%20Conf.%20on%20Genetic%20Algorithms,%201991) is from 1991 not 1989
- Please rationalise the text for 'Proceedings of the 5th International Conference on Genetic Algorithms' with the one for the 4th. They should not look so different when they are the same conference two y ears apart.
