### 2024-10-17 0de1b9e
fix ICGA references and formatting issues.
