+++ paper/paper.bib
@@ -23,7 +23,7 @@
   title     = {Adaptation in Natural and Artificial Systems},
   author    = {Holland, John H.},
   year      = {1975},
-  publisher = {University of Michigan Press},
+  publisher = {MIT Press},
   address   = {Ann Arbor, MI},
   isbn      = {9780262581110}
 }
@@ -49,13 +49,26 @@
   doi       = {10.1016/b978-0-08-050684-5.50008-2}
 }
 
-@inproceedings{manderick1989genetic,
+@inproceedings{manderick1991genetic,
   title       = {The Genetic Algorithm and the Structure of the Fitness Landscape},
   author      = {Manderick, Bernard and Spiessens, Peter},
-  booktitle   = {Proceedings of the 4th International Conference on Genetic Algorithms},
+  booktitle   = {Proceedings of the 4th International Conference on Genetic Algorithms (ICGA)},
+  editor      = {Belew, Richard K. and Booker, Lashon B.}, 
   pages       = {143--150},
-  year        = {1989},
-  organization = {Morgan Kaufmann Publishers Inc.}
+  year        = {1991},
+  publisher   = {Morgan Kaufmann},
+  address     = {San Diego, CA, USA}
+}
+
+@inproceedings{whitley1993cellular,
+  title     = {Cellular Genetic Algorithms},
+  author    = {Whitley, Darrell},
+  booktitle = {Proceedings of the 5th International Conference on Genetic Algorithms (ICGA)},
+  editor    = {Forrest, Stephanie},
+  pages     = {658},
+  year      = {1993},
+  publisher = {Morgan Kaufmann},
+  address   = {San Mateo, CA, USA}
 }
 
 @book{michalewicz1996genetic,
@@ -79,16 +92,6 @@
   doi       = {10.1007/978-3-662-44874-8}
 }
 
-@inproceedings{whitley1993cellular,
-  title     = {Cellular genetic algorithms},
-  author    = {Whitley, Darrell},
-  booktitle = {Proceedings of the Fifth International Conference on Genetic Algorithms (ICGA)},
-  editor    = {Forrest, Stephanie},
-  pages     = {658},
-  year      = {1993},
-  publisher = {Morgan Kaufmann},
-  address   = {California, CA, USA}
-}
 
 @book{alba2008cellular,
   title     = {Cellular Genetic Algorithms},
