### 2024-10-17 1ab10a5
fix missing references in paper.md
