### 2024-10-17 #7322 @SevgiAkten
Hello @JBorrow, thank you for the detailed feedback.

I’ve implemented the requested changes and ensured the ICGA references are now consistent.
