### 2024-10-31 SevgiAkten/pycellga#65 @jbussemaker
JOSS Review Feedback
Part of the JOSS paper review: https://github.com/openjournals/joss-reviews/issues/7322

Hi @SevgiAkten, I have taken an initial look at pycellga and I have some questions and feedback:
- [x] What does "machine-coded" mean? Does it simply mean that they are programmed (aka written in code)? If that is the case, this is not something special, because all GA implementations are code so you cannot not machine-code the operators...
- [x] To improve reproducability (for example in tutorials or in the usage example in the readme/paper), you may introduce a "seed" parameter, so that even with randomness you always get the same result
- [x] The documentation available at https://sevgiakten.github.io/pycellga/ is severly lacking: currently it is only an API reference of the Python package (however please remove the documentation of the tests), which is also important, but usually documentation also includes things like an introduction (what/why, installation, citing instructions, contribution guidelines, etc.), tutorials (a good idea here is to use Jupyter notebooks), and examples. See for example https://cosapp.readthedocs.io/ or https://sbarchopt.readthedocs.io/, and also have a look at the "Documentation" section of our reviewer checklists above.
- [x] It's good to see that you're using CI/CD, however you can consider:
  - [x] Running automated unit tests
  - [x] Managing releases in Github and use CI/CD to publish the package to PyPI
- [x] Your optimization problem configuration is distributed between the problem class and the optimizer parameters. Please move all problem-related settings, like design variables (type and bounds), objectives, constraints, and the evaluation function to the problem class. You can create a base class from which all problem classes inherit to "force" the user to adhere to a template. Also see the comment below.
- [x] For comparison to other evolutionary algorithm approaches, consider interfacing with popular packages like [pymoo](https://pymoo.org/). This would also open up access to many more test problems. You can even consider inheriting from pymoo more, for example by reusing its `Problem` class (and associated classes like design variable types) to define optimization problems.
- [x] Consider returning results of your optimization as a class (maybe a dataclass could be useful?) instead of a list of variables, to make it easier to understand the outputs
- [x] Please switch to the more modern `setup.cfg` with a `pyproject.toml` file. For more info see https://godatadriven.com/blog/a-practical-guide-to-setuptools-and-pyproject-toml/
- [x] Please use absolute module imports, for example replace `from mutation.mutation_operator import ...` with `from pycellga.mutation.mutation_operator import ...`. Relative imports may work in your local dev environment, but may conflict with other packages in a larger Python environment.
- [x] The paper includes a "statement of need" section, but the statement is pretty generic: it mentions complexity and limitations of traditional GA, however not exactly how these are tackled. Also it is not clear what complexity means (nr of design variables, types of design variables, mixed types, nr of objectives, constraints, non-linearity?), and why traditional GA has limitations regarding this. Also, traditional GA has a huge body of research tackling all kinds of issues with earlier GA's, so there has to be some research that tackles the same limitations that CGA tries to tackle: please show this in the paper.

Further review is in progress.
