### 2024-11-04 SevgiAkten/pycellga#66 @SevgiAkten
Clarification on "Machine-coded" Terminology
This is linked to the review feedback in Issue #65 and part of the JOSS review: openjournals/joss-reviews#7322
