### 2024-11-04 47779a6
(#66) Clarified "machine-coded" terminology in documentation
