+++ paper/paper.md
@@ -35,7 +35,11 @@ Despite their effectiveness, traditional GAs face challenges, particularly in ma
 
 `pycellga` is a Python package designed to efficiently implement CGAs. By integrating the principles of cellular automata with genetic algorithms, `pycellga` offers a robust framework for tackling complex optimization problems. The `pycellga` package is designed to handle a wide range of optimization problems, including binary, real-valued, and permutation-based challenges, making it a versatile tool for diverse applications in evolutionary computation. The package includes several built-in functions for initialization, selection, crossover, mutation, and evaluation, as well as customization options to cater to different needs. This flexibility allows researchers and practitioners to apply CGAs to a wide range of problems with ease [@karakaya2024improved].
 
-By providing a comprehensive toolkit for CGAs, `pycellga` aims to advance the field of evolutionary computation and equip researchers with the tools needed to solve increasingly complex optimization problems effectively. The integration of cellular automata with genetic algorithms in `pycellga` represents a significant advancement, offering greater flexibility and adaptability compared to traditional methods [@michalewicz1996genetic; @eiben2003introduction; @karakaya2024improved]. `pycellga` has machine coded operators with byte implementations, developed by [@satman2013machine]. Additionally, it features Alpha-male CGA, developed based on insights from [@satman2019alpha_male_ga], Machine-Coded Compact CGA, developed based on insights from [@satman2020machine_coded_cga], and Improved CGA with Machine-Coded Operators [@karakaya2024improved]. The improved cellular genetic algorithm uses machine-coded operators specifically designed for real-valued optimization problems. This method stands out by employing byte-based operators, which are crafted to efficiently process numerical data in terms of memory usage.
+By providing a comprehensive toolkit for CGAs, `pycellga` aims to advance the field of evolutionary computation and equip researchers with the tools needed to solve increasingly complex optimization problems effectively. The integration of cellular automata with genetic algorithms in `pycellga` represents a significant advancement, offering greater flexibility and adaptability compared to traditional methods [@michalewicz1996genetic; @eiben2003introduction; @karakaya2024improved]. 
+
+The `pycellga` package includes machine-coded operators with byte-level implementations, developed by [@satman2013machine]. In the context of genetic algorithms, "machine-coded" refers to a specialized encoding technique optimized for real-parameter optimization. This approach differs from standard coding practices by emphasizing efficient data processing through byte-level manipulation. Originally introduced by [@satman2013machine], this technique is particularly advantageous for real-valued optimization tasks, as it allows direct manipulation of byte-representations to enhance computational performance. Encoding and decoding of numerical values conform to the IEEE 754 standard for floating-point arithmetic, further improving precision and effectiveness in optimizing continuous functions. By using machine-coded operators, `pycellga` leverages this efficiency to handle complex optimization challenges more effectively.
+
+In addition, the `pycellga` package features Alpha-male CGA, developed based on insights from [@satman2019alpha_male_ga]; Machine-Coded Compact CGA, inspired by [@satman2020machine_coded_cga]; and Improved CGA with Machine-Coded Operators [@karakaya2024improved]. The improved cellular genetic algorithm utilizes machine-coded operators specifically tailored for real-valued optimization problems. This method is particularly distinctive for its use of byte-based operators, which are designed to process numerical data efficiently in terms of memory usage.
 
 
 # State of the field
