### 2024-11-04 SevgiAkten/pycellga#66 @SevgiAkten
Reviewer Comment Response

Hi @jbussemaker,

Thank you for your feedback regarding the term "machine-coded." We have clarified this terminology in the documentation to provide a more precise understanding. In this context, "machine-coded" refers to a specialized encoding technique optimized for real-parameter optimization, originally introduced by Satman (2013). This approach focuses on efficient data processing through byte-level manipulation, which is particularly advantageous for real-valued optimization tasks. We have also included a reference to the IEEE 754 standard for floating-point arithmetic, further enhancing precision and effectiveness in optimizing continuous functions.

Please let us know if this update aligns with your expectations or if further clarification is needed.
