### 2024-11-04 b2dba52
(#67) Added seed parameter for reproducibility
