### 2024-11-04 SevgiAkten/pycellga#66 @jbussemaker
All right, if you indeed make it clear that "machine-coded" is actually the name of some encoding technique that you reference to, then for the paper it is fine.

However just for my understanding: there have been many efforts over the past decades to support real-valued design variables in evolutionary algorithms, mostly through the development of specialized crossover and mutation operators (instead of relying on operators developed for byte strings). Why don't you use these well-tested operators?
