### 2024-11-04 SevgiAkten/pycellga#66 @SevgiAkten
We decided to implement machine-coded, byte-based operators due to their effectiveness in handling real-valued optimization tasks. Byte-level manipulations provide computational efficiency and have shown promising results in finding optimal solutions compared to traditional floating-point operators. In our previous study, we demonstrated these advantages with an Improved Cellular Genetic Algorithm using machine-coded operators, showing better performance on benchmark functions like Ackley, Bohachevsky, and Rastrigin (Akten Karakaya & Satman, 2024).

As part of my doctoral dissertation, I expanded the test benchmarks to include 20 functions, where machine-coded CGAs outperformed classical real-valued operators in 16 cases, including complex functions like Bent Cigar and Zakharov (Akten Karakaya, 2024). To give users more flexibility, pycellga also includes traditional operators, such as Arithmetic, Linear, and BLX-alpha crossover, allowing for comparisons within the same framework.

References:

[Akten Karakaya, S., & Satman, M. H. (2024). An improved cellular genetic algorithm with machine-coded operators for real-valued optimization problems. Journal of Engineering Research and Applied Science, 13(1), 2500-2514.](http://www.journaleras.com/index.php/jeras/article/view/334)

[Akten Karakaya, S. (2024). Investigation of cellular genetic algorithms and improvement of their performance (Doctoral dissertation, İstanbul University). Council of Higher Education Thesis Center (YÖKTEZ), Turkey.](https://tez.yok.gov.tr/UlusalTezMerkezi/giris.jsp)
