### 2024-11-04 SevgiAkten/pycellga#68 @SevgiAkten
Enhance Documentation
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
