### 2024-11-04 SevgiAkten/pycellga#66 @jbussemaker
Thank you for the explanation. For the JOSS review this is sufficient, especially because you also include the traditional operators for comparison.
