### 2024-11-05 379005c
(#67) edit seed_par
