+++ pycellga/example/example_sync_cga.py
@@ -64,7 +64,8 @@ def run_sync_cga_example():
         recombination=optimizer.BlxalphaCrossover,
         mutation=optimizer.FloatUniformMutation,
         mins=[-32.768] * 5,  # Minimum values for each gene
-        maxs=[32.768] * 5    # Maximum values for each gene
+        maxs=[32.768] * 5,    # Maximum values for each gene
+        seed_par=100
     )
 
     # Print the results
