### 2024-11-06 2e240b3
(#68) Enhance doc structure
