+++ docs/modules.rst
@@ -1,7 +1,0 @@
-pycellga
-========
-
-.. toctree::
-   :maxdepth: 4
-
-   pycellga
