+++ docs/pycellga.problems.single_objective.discrete.permutation.rst
@@ -1,21 +1,16 @@
-pycellga.problems.single\_objective.discrete.permutation package
-================================================================
+Permutation-Based Optimization Problems
+===================================================
 
-Submodules
-----------
+The `pycellga.problems.single_objective.discrete.permutation` package provides a set of benchmark problems that involve permutation-based solutions. These problems are particularly useful in evaluating optimization algorithms designed for sequencing and ordering tasks, where the solution is represented as a permutation.
 
-pycellga.problems.single\_objective.discrete.permutation.tsp module
--------------------------------------------------------------------
+
+**Traveling Salesman Problem (TSP)**
+--------------
+
+The Traveling Salesman Problem (TSP) is a classic optimization problem where the goal is to find the shortest possible route that visits each city exactly once and returns to the origin city. TSP is widely used to evaluate the efficiency of algorithms that handle combinatorial and sequencing problems. This module provides a standard implementation of TSP for testing purposes.
 
 .. automodule:: pycellga.problems.single_objective.discrete.permutation.tsp
    :members:
    :undoc-members:
    :show-inheritance:
 
-Module contents
----------------
-
-.. automodule:: pycellga.problems.single_objective.discrete.permutation
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
