### 2024-11-06 SevgiAkten/pycellga#69 @SevgiAkten
Set Up CI/CD for Automated Testing and PyPI Deployment
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
