+++ .github/workflows/documentation.yml
@@ -3,7 +3,8 @@ name: Build and Deploy Docs
 on:
   push:
     branches:
-      - main  
+      - main
+  workflow_dispatch:
 
 jobs:
   build:
