### 2024-11-07 0c4d77e
( #70) Add PyPI badge and project URLs
