### 2024-11-07 SevgiAkten/pycellga#71 @SevgiAkten
Centralize Problem Configuration and Integrate pymoo Compatibility
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
