### 2024-11-12 8ae1f55
(#71) Consolidate problem settings in base class and integrate pymoo compatibility
