+++ pycellga/tests/test_fms.py
@@ -1,7 +1,6 @@
 import pytest
-from problems.abstract_problem import AbstractProblem
-from problems.single_objective.continuous.fms import Fms  # Replace with the actual path if different
-from numpy import random
+from problems.single_objective.continuous.fms import Fms 
+import numpy as np
 
 @pytest.fixture
 def fms_instance():
@@ -21,8 +20,8 @@ def test_fms(fms_instance):
     """
     Test the Fms function implementation.
 
-    This test checks the calculation of the FMS function value for a given list of binary variables.
-    It uses a predefined input and compares the output to the expected value.
+    This test checks the calculation of the Fms function value for a given list of float variables.
+    It uses a predefined input and compares the output to ensure it's in the correct format.
 
     Parameters
     ----------
@@ -31,31 +30,25 @@ def test_fms(fms_instance):
 
     Notes
     -----
-    The test uses a randomly generated sample input chromosome of length 192 and checks if the function output is a float
-    and non-negative. Additional checks with known values can be added for more thorough testing.
+    The test generates a sample input chromosome of length 6 within the valid bounds and checks:
+    - If the function output is a float.
+    - If the fitness value is non-negative, as it represents a sum of squares of differences.
 
     Assertions
     ----------
     - The fitness value should be a float.
-    - The fitness value should be non-negative, as it's a sum of squares of differences.
-
-    Examples
-    --------
-    >>> test_fms(fms_instance)
+    - The fitness value should be non-negative.
     """
-    # Define a sample input chromosome (binary list)
-    sample_chromosome = [random.randint(2) for _ in range(192)]
+    # Define a sample input chromosome within bounds [-6.4, 6.35] for each variable
+    sample_chromosome = np.random.uniform(-6.4, 6.35, size=6)
 
     # Calculate the FMS function value for the sample input
     fitness_value = fms_instance.f(sample_chromosome)
 
     # Assertions to check correctness
-    assert isinstance(fitness_value, float)
-    assert fitness_value >= 0  # Since it's a sum of squares of differences
-
-    # Additional checks with known values
-    # Here we assume specific values and their known outputs for more thorough testing
-    # You can add more specific test cases if you have known outputs for certain inputs
+    assert isinstance(fitness_value, float), "Fitness value should be a float."
+    assert fitness_value >= 0, "Fitness value should be non-negative."
+    print(f"Sample chromosome: {sample_chromosome}, Fitness: {fitness_value}")
 
 if __name__ == "__main__":
     pytest.main()
