+++ pycellga/problems/single_objective/continuous/rastrigin.py
@@ -1,15 +1,21 @@
 from numpy import cos, pi
 from problems.abstract_problem import AbstractProblem
+
 class Rastrigin(AbstractProblem):
     """
     Rastrigin function implementation for optimization problems.
 
     The Rastrigin function is widely used for testing optimization algorithms.
-    The function is usually evaluated on the hypercube x_i ∈ [-5.12, 5.12], for all i = 1, 2, ..., n.
+    It is typically evaluated on the hypercube x_i ∈ [-5.12, 5.12], for all i = 1, 2, ..., n.
 
     Attributes
     ----------
-    None
+    design_variables : int
+        The number of variables for the problem.
+    bounds : list of tuple
+        The bounds for each variable, typically [(-5.12, 5.12), (-5.12, 5.12), ...].
+    objectives : int
+        Number of objectives, set to 1 for single-objective optimization.
 
     Methods
     -------
@@ -21,6 +27,19 @@ class Rastrigin(AbstractProblem):
     -5.12 ≤ xi ≤ 5.12 for i = 1,…,n
     Global minimum at f(0,...,0) = 0
     """
+
+    def __init__(self, design_variables=2):
+        """
+        Initialize the Rastrigin problem with the specified number of variables.
+
+        Parameters
+        ----------
+        design_variables : int, optional
+            The number of design variables, by default 2.
+        """
+        self.design_variables = design_variables
+        self.bounds = [(-5.12, 5.12) for _ in range(design_variables)]
+        self.objectives = 1
 
     def f(self, x: list) -> float:
         """
@@ -36,5 +55,9 @@ class Rastrigin(AbstractProblem):
         float
             The Rastrigin function value.
         """
+        if len(x) != self.design_variables:
+            raise ValueError(f"Input must have exactly {self.design_variables} variables.")
+
         A = 10.0
-        return round((A * len(x)) + sum([(i * i) - (A * cos(2 * pi * i)) for i in x]), 3)
+        fitness = (A * self.design_variables) + sum([(xi ** 2) - (A * cos(2 * pi * xi)) for xi in x])
+        return round(fitness, 3)
