### 2024-11-12 SevgiAkten/pycellga#72 @SevgiAkten
Refactor Optimization Results to Use a Data Class for Improved Clarity and Structure
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
