### 2024-11-12 2f9083b
(#72) Refactor results and tests for dataclass structure
