+++ pycellga/example/example_ccga.py
@@ -43,12 +43,14 @@ def run_ccga_example():
     
     Returns
     -------
-    tuple
-        A tuple containing the best solution chromosome and its corresponding value.
+    Result
+        A Result instance containing the best solution's chromosome, its fitness value, 
+        and the generation in which it was found.
     """
     # Create an instance of the problem
     problem_instance = ExampleProblem()
 
+    # Run the optimizer and get the result
     result = optimizer.ccga(
         n_cols=5,
         n_rows=5,
@@ -62,7 +64,9 @@ def run_ccga_example():
     )
 
     # Print the results
-    print("Best solution:", result[1], "\nBest solution chromosome:", result[0])
+    print("Best solution chromosome:", result.chromosome)
+    print("Best fitness value:", result.fitness_value)
+    print("Generation found:", result.generation_found)
 
 if __name__ == "__main__":
     run_ccga_example()
