### 2024-11-12 086d271
Bump version to 0.2.2 for new release
