+++ setup.py
@@ -5,7 +5,7 @@ with open('requirements.txt') as f:
 
 setup(
     name='pycellga',  
-    version='0.2.1',  
+    version='0.2.2',  
     description='A Python Package for Improved Cellular Genetic Algorithms',  
     author='Sevgi Akten Karakaya, Mehmet Hakan Satman',  
     author_email='sevgiakten@gmail.com, mhsatman@gmail.com',  
