### 2024-11-13 SevgiAkten/pycellga#73 @SevgiAkten
Switch to Modern File Structure: setup.cfg and pyproject.toml
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
