### 2024-11-15 12e0065
(#73) add setup.cfg and pyproject.toml files
