+++ setup.cfg
@@ -1,0 +1,59 @@
+[metadata]
+name = pycellga
+version = 0.2.3
+description = A Python Package for Improved Cellular Genetic Algorithms
+author = Sevgi Akten Karakaya, Mehmet Hakan Satman
+author_email = sevgiakten@gmail.com, mhsatman@gmail.com
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/SevgiAkten/pycellga
+project_urls =
+    Documentation = https://sevgiakten.github.io/pycellga/
+    Source = https://github.com/SevgiAkten/pycellga
+    Issue Tracker = https://github.com/SevgiAkten/pycellga/issues
+    PyPI = https://pypi.org/project/pycellga/
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
+python_requires = >=3.6
+install_requires =
+    geographiclib==2.0
+    geopy==2.4.1
+    grip==4.6.2
+    iniconfig==2.0.0
+    Markdown==3.6
+    Markdown-Editor==1.0.7
+    mpmath==1.3.0
+    numpy==1.26.2
+    path-and-address==2.0.1
+    pluggy==1.5.0
+    plumbum==1.8.3
+    ply==3.11
+    pyparsing==3.1.1
+    pytest==8.3.2
+    pytest-xdist==3.6.1
+    python-dateutil==2.8.2
+    pytz==2023.3.post1
+    six==1.16.0
+    Sphinx==7.4.7
+    sphinx-rtd-theme==2.0.0
+    sphinxcontrib-applehelp==2.0.0
+    sphinxcontrib-devhelp==2.0.0
+    sphinxcontrib-htmlhelp==2.1.0
+    sphinxcontrib-jquery==4.1
+    sphinxcontrib-jsmath==1.0.1
+    sphinxcontrib-qthelp==2.0.0
+    sphinxcontrib-serializinghtml==2.0.0
+    tsplib95==0.7.1
+    tzdata==2023.3
+    pymoo==0.5.0
+
+[options.packages.find]
+where = .
+
+[options.package_data]
+* = *.txt, *.md
