### 2024-11-15 SevgiAkten/pycellga#74 @SevgiAkten
Switch from Relative to Absolute Imports for Robust Package Integration
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
