+++ pycellga/tests/test_insertion_mutation.py
@@ -1,6 +1,6 @@
-from mutation.insertion_mutation import InsertionMutation
-from problems.single_objective.discrete.permutation.tsp import Tsp
-from individual import Individual, GeneType
+from pycellga.mutation.insertion_mutation import InsertionMutation
+from pycellga.problems.single_objective.discrete.permutation.tsp import Tsp
+from pycellga.individual import Individual, GeneType
 import random
 
 def test_insertion_mutation():
