+++ pycellga/problems/single_objective/discrete/binary/maxcut20_01.py
@@ -1,4 +1,4 @@
-from problems.abstract_problem import AbstractProblem
+from pycellga.problems.abstract_problem import AbstractProblem
 
 class Maxcut20_01(AbstractProblem):
     """
