+++ pycellga/mutation/bit_flip_mutation.py
@@ -1,7 +1,7 @@
 import numpy as np
 from individual import *
-from problems.abstract_problem import AbstractProblem
-from mutation.mutation_operator import MutationOperator
+from pycellga.problems.abstract_problem import AbstractProblem
+from pycellga.mutation.mutation_operator import MutationOperator
 
 class BitFlipMutation(MutationOperator):
     """
