+++ pycellga/problems/single_objective/continuous/levy.py
@@ -1,4 +1,4 @@
-from problems.abstract_problem import AbstractProblem
+from pycellga.problems.abstract_problem import AbstractProblem
 import math
 from mpmath import power as pw
 from typing import List
