+++ pycellga/recombination/pmx_crossover.py
@@ -1,7 +1,7 @@
-from individual import *
-from problems.abstract_problem import AbstractProblem
+from pycellga.individual import *
+from pycellga.problems.abstract_problem import AbstractProblem
 from typing import List
-from recombination.recombination_operator import RecombinationOperator
+from pycellga.recombination.recombination_operator import RecombinationOperator
 
 class PMXCrossover(RecombinationOperator):
     """
