+++ pycellga/tests/test_zakharov_function.py
@@ -1,5 +1,5 @@
 import pytest
-from problems.single_objective.continuous.zakharov import Zakharov
+from pycellga.problems.single_objective.continuous.zakharov import Zakharov
 
 @pytest.fixture
 def setup_zakharov():
