### 2024-11-15 7ab1cb5
(#74) Bump version to 0.2.4
