+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pycellga
-version = 0.2.3
+version = 0.2.4
 description = A Python Package for Improved Cellular Genetic Algorithms
 author = Sevgi Akten Karakaya, Mehmet Hakan Satman
 author_email = sevgiakten@gmail.com, mhsatman@gmail.com
