### 2024-11-18 SevgiAkten/pycellga#75 @SevgiAkten
Enhance 'Statement of Need' with Specificity and Supporting Literature
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
