+++ paper/paper.bib
@@ -125,3 +125,13 @@
   year      = {2020},
   doi       = {10.17093/alphanumeric.576919}
 }
+
+@book{coello2007multiobjective,
+  author    = {Carlos A. Coello Coello, Gary B. Lamont, and David A. Van Veldhuizen},
+  title     = {Evolutionary Algorithms for Solving Multi-Objective Problems},
+  publisher = {Springer},
+  year      = {2007},
+  edition   = {2nd},
+  isbn      = {978-0387332542},
+  doi       = {10.1007/978-0-387-36797-2}
+}
