+++ paper/paper.md
@@ -49,7 +49,16 @@ There are several existing software packages that implement genetic algorithms, 
 
 # Statement of need 
 
-The need for `pycellga` arises from the increasing complexity of optimization problems and the limitations of traditional genetic algorithms in handling these complexities. By incorporating cellular automata, `pycellga` introduces localized interactions and diversity within the population, which can lead to more effective and efficient solutions. This package is particularly useful for researchers and practitioners who wish to explore advanced genetic algorithm techniques. `pycellga` includes machine-coded operators with byte-level implementations developed by [@satman2013machine]. Additionally, it features Alpha-male CGA, Machine-Coded Compact CGA, and an Improved CGA with Machine-Coded Operators for real-valued optimization problems [@karakaya2024improved].
+The need for advanced optimization frameworks like `pycellga` arises from the growing complexity of real-world problems, which often involve high-dimensional search spaces, mixed-variable types, and intricate constraints. Traditional Genetic Algorithms (GAs), while highly versatile and widely adopted, face several limitations in addressing these challenges. These include a tendency toward premature convergence, difficulty in maintaining population diversity, and an inability to balance global exploration with local exploitation effectively [@holland1975adaptation; @goldberg1989genetic].
+
+Cellular Genetic Algorithms (CGAs) offer a compelling solution by introducing a spatially structured population. In CGAs, individuals are arranged in a grid-like topology and interact only with their neighbors. This localized interaction not only enhances population diversity but also mitigates the risk of premature convergence, a common issue in traditional GAs [@manderick1991genetic; @whitley1993cellular]. By promoting a gradual spread of genetic material through localized selection and crossover, CGAs achieve a more thorough exploration of the solution space, particularly in high-dimensional and multi-modal optimization problems [@alba2008cellular]. These attributes make CGAs especially effective for tackling complex challenges in scheduling, resource allocation, and multi-objective optimization [@coello2007multiobjective].
+
+Despite the clear advantages of CGAs, their implementation in existing tools remains limited. Frameworks such as JCell, while functional, lack the flexibility, extensibility, and user-friendly features offered by modern programming environments like Python [@alba2008cellular]. To address this gap, `pycellga` provides a Python-based framework that combines the strengths of CGAs with the convenience and power of Python's ecosystem. The package includes efficient byte-level implementations of machine-coded operators, as introduced by [@satman2013machine], which significantly enhance performance for real-valued optimization problems. Additionally, it incorporates advanced CGA variants, such as Alpha-male CGA [@satman2019alpha_male_ga], Machine-Coded Compact CGA [@satman2020machine_coded_cga], and an Improved CGA with Machine-Coded Operators [@karakaya2024improved].
+
+The unique contributions of `pycellga` extend beyond its robust implementation of CGAs. By supporting binary, real-valued, and permutation-based optimization problems, the package offers unparalleled versatility. Its use of machine-coded operators for real-valued optimization, adhering to IEEE 754 floating-point arithmetic standards, ensures high precision and computational efficiency. Moreover, `pycellga` is designed to be extensible, enabling users to easily customize selection, crossover, and mutation operators to suit specific problem domains.
+
+In a field where traditional GAs have seen extensive research to address their inherent limitations [@eiben2003introduction], CGAs provide a novel and effective approach to further these advancements. `pycellga` stands out as a modern, accessible, and feature-rich toolkit that equips researchers and practitioners to tackle increasingly complex optimization challenges with confidence.
+
 
 
 # Installation and basic usage
