### 2024-11-25 #7322 @jmejia8
Dear @SevgiAkten,
The following are my comments.

Regarding the paper:

- [ ] It is not clear for me whether CGA is a contribution from this paper and software, the confusion raises due to lines 16-17. This is later clarified in Section "Statement of need" (line 107), but it could be added in "Summary" as well.

> The integration of cellular automata with genetic algorithms in pycellga represents a significant advancement in the field of evolutionary computation, offering increased flexibility and adaptability compared to traditional methods

- [ ] I suggest to the authors to discuss why other libraries in Python (and other languages) are not including CGA variants in their list of implemented algorithms, considering that "... The integration of cellular automata with genetic algorithms in pycellga represents a significant advancement". This could be addressed in Section "State of the field"

Regarding the software:

- [ ] Experiments and installation proceed as stated in paper.
- [ ] The documentation is clear and with some illustrative examples. However, no examples are given to customize the methods and operators.
- [ ] Installation is intuitive via PIP package manager, but slow. Perhaps  `pycellga` has many requirements (dependencies), and it seems that many of them are not strictly necessary (e.g., `geographiclib, geopy, pymoo`, among others). Could you verify or clarify this?
- [ ] In general, the codebase is well-structured and documented, extensible and easy to understand. I do not have any concern on this.
