### 2024-11-26 SevgiAkten/pycellga#76 @SevgiAkten
Clarify CGA's Role in Summary Section to Address Reader Confusion
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
