+++ paper/paper.md
@@ -25,7 +25,12 @@ bibliography: paper.bib
 # Summary
 `pycellga` is a Python package that implements cellular genetic algorithms (CGAs) for optimizing complex problems. CGAs combine the principles of cellular automata and traditional genetic algorithms, utilizing a spatially structured population organized in a grid-like topology. This structure allows each individual to interact only with its neighboring individuals, promoting diversity and maintaining a balance between exploration and exploitation during the optimization process.
 
-The package is designed to be user-friendly, with a straightforward installation process and comprehensive documentation. Researchers and practitioners in fields such as operations research, artificial intelligence, and machine learning can leverage `pycellga` to tackle complex optimization challenges effectively. The integration of cellular automata with genetic algorithms in `pycellga` represents a significant advancement in the field of evolutionary computation, offering increased flexibility and adaptability compared to traditional methods. `pycellga` also includes machine-coded operators with byte implementations, developed by [@satman2013machine]. Additionally, it features Alpha-male CGA, Machine-Coded Compact CGA, and Improved CGA with Machine-Coded Operators for real-valued optimization problems [@karakaya2024improved].
+While CGAs themselves are not a novel contribution of this work, `pycellga` significantly enhances their applicability by integrating advanced features and providing unparalleled versatility. The package supports binary, real-valued, and permutation-based optimization problems, making it adaptable to a wide variety of problem domains. Its use of machine-coded operators for real-valued optimization, adhering to IEEE 754 floating-point arithmetic standards, ensures high precision and computational efficiency. Moreover, `pycellga` is designed to be extensible, enabling users to easily customize selection, crossover, and mutation operators to suit specific problem requirements.
+
+The package is designed to be user-friendly, with a straightforward installation process and comprehensive documentation. Researchers and practitioners in fields such as operations research, artificial intelligence, and machine learning can leverage `pycellga` to tackle complex optimization challenges effectively. By integrating the principles of cellular automata with genetic algorithms, `pycellga` represents a significant advancement in the field of evolutionary computation, offering increased flexibility and adaptability compared to traditional methods.
+
+Additionally, `pycellga` includes machine-coded operators with byte implementations, developed by [@satman2013machine]. It features Alpha-male CGA, Machine-Coded Compact CGA, and Improved CGA with Machine-Coded Operators for real-valued optimization problems [@karakaya2024improved].
+
 
 # Introduction
 
