### 2024-11-26 SevgiAkten/pycellga#77 @SevgiAkten
Discuss the Lack of CGA Variants in Existing Libraries in the 'State of the Field' Section
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
