### 2024-11-26 7b4c9cf
(#77) Address CGA Gaps in Existing Libraries
