+++ pycellga/tests/test_arithmetic_crossover.py
@@ -1,8 +1,8 @@
 import pytest
 import random
-from pycellga.individual import Individual, GeneType
-from pycellga.problems.abstract_problem import AbstractProblem
-from pycellga.recombination.arithmetic_crossover import ArithmeticCrossover 
+from individual import Individual, GeneType
+from problems.abstract_problem import AbstractProblem
+from recombination.arithmetic_crossover import ArithmeticCrossover 
 
 class MockProblem(AbstractProblem):
     """
