+++ pycellga/problems/single_objective/continuous/rothellipsoid.py
@@ -1,4 +1,4 @@
-from pycellga.problems.abstract_problem import AbstractProblem
+from problems.abstract_problem import AbstractProblem
 from mpmath import power as pw
 
 class Rothellipsoid(AbstractProblem):
