+++ pycellga/tests/test_two_opt_mutation.py
@@ -1,6 +1,6 @@
-from pycellga.mutation.two_opt_mutation import TwoOptMutation
-from pycellga.problems.single_objective.discrete.permutation.tsp import Tsp
-from pycellga.individual import Individual, GeneType
+from mutation.two_opt_mutation import TwoOptMutation
+from problems.single_objective.discrete.permutation.tsp import Tsp
+from individual import Individual, GeneType
 import random
 
 def test_two_opt_mutation():
