+++ pycellga/problems/single_objective/continuous/griewank.py
@@ -1,4 +1,4 @@
-from pycellga.problems.abstract_problem import AbstractProblem
+from problems.abstract_problem import AbstractProblem
 import math
 from typing import List
 
