+++ pycellga/recombination/pmx_crossover.py
@@ -1,7 +1,7 @@
-from pycellga.individual import *
-from pycellga.problems.abstract_problem import AbstractProblem
+from individual import *
+from problems.abstract_problem import AbstractProblem
 from typing import List
-from pycellga.recombination.recombination_operator import RecombinationOperator
+from recombination.recombination_operator import RecombinationOperator
 
 class PMXCrossover(RecombinationOperator):
     """
