### 2024-11-28 SevgiAkten/pycellga#78 @SevgiAkten
Enhance Documentation with Customization Scenarios
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
