### 2024-11-28 fe23c2d
(#78) Add customization examples to docs
