+++ _sources/usage_examples.rst.txt
@@ -63,3 +63,204 @@ Usage
 We have provided a basic example above. If you're interested in exploring more examples:
 
 - `Click here to see the other examples <https://github.com/SevgiAkten/pycellga/tree/main/pycellga/example>`_ available directly in the repository.
+
+
+Customization Scenarios
+------------------------
+
+`pycellga` allows users to customize the methods and operators used in their optimization problems. Below are some examples to help you design your own scenarios.
+
+Selecting Gene Types
+---------------------
+
+You can change the `gen_type` parameter based on the type of problem:
+
+- `pycellga.optimizer.GeneType.REAL` for real-valued problems.
+- `pycellga.optimizer.GeneType.BINARY` for binary problems.
+- `pycellga.optimizer.GeneType.PERMUTATION` for permutation-based problems.
+
+
+Problem Selection
+------------------
+
+`pycellga` includes several built-in problems for quick usage, categorized into continuous (real-valued), binary, and permutation problems. These problems allow users to directly specify the `problem` parameter for quick implementation.
+
+Continuous Problems
+--------------------
+
+For continuous optimization problems, the following functions are available:
+
+- `pycellga.optimizer.Ackley()`
+- `pycellga.optimizer.Bentcigar()`
+- `pycellga.optimizer.Bohachevsky()`
+- `pycellga.optimizer.Chichinadze()`
+- `pycellga.optimizer.Dropwave()`
+- `pycellga.optimizer.Fms()`
+- `pycellga.optimizer.Griewank()`
+- `pycellga.optimizer.Holzman()`
+- `pycellga.optimizer.Levy()`
+- `pycellga.optimizer.Matyas()`
+- `pycellga.optimizer.Pow()`
+- `pycellga.optimizer.Powell()`
+- `pycellga.optimizer.Rastrigin()`
+- `pycellga.optimizer.Rosenbrock()`
+- `pycellga.optimizer.Rothellipsoid()`
+- `pycellga.optimizer.Schaffer()`
+- `pycellga.optimizer.Schaffer2()`
+- `pycellga.optimizer.Schwefel()`
+- `pycellga.optimizer.Sphere()`
+- `pycellga.optimizer.StyblinskiTang()`
+- `pycellga.optimizer.Sumofdifferentpowers()`
+- `pycellga.optimizer.Threehumps()`
+- `pycellga.optimizer.Zakharov()`
+- `pycellga.optimizer.Zettle()`
+
+Binary Problems
+----------------
+
+For binary optimization problems, use the following built-in options:
+
+- `pycellga.optimizer.CountSat()`
+- `pycellga.optimizer.Ecc()`
+- `pycellga.optimizer.Maxcut20_01()`
+- `pycellga.optimizer.Maxcut20_09()`
+- `pycellga.optimizer.Maxcut100()`
+- `pycellga.optimizer.Mmdp()`
+- `pycellga.optimizer.OneMax()`
+- `pycellga.optimizer.Peak()`
+
+Permutation Problems
+--------------------
+
+For permutation-based optimization problems, the following option is available:
+
+- `pycellga.optimizer.Tsp()`
+
+These built-in problems provide a diverse set of test cases, allowing users to explore `pycellga`'s capabilities across a wide range of optimization challenges. Users can also define custom problems to suit their specific needs.
+
+
+Selection Operators
+--------------------
+Choose from a variety of selection methods:
+
+- `pycellga.optimizer.TournamentSelection`
+- `pycellga.optimizer.RouletteWheelSelection`
+
+Recombination Operators
+------------------------
+
+The package provides multiple crossover operators:
+
+- `pycellga.optimizer.OnePointCrossover`
[…]
