### 2024-11-28 2e7e4e8
Bump version to 0.3.0
