### 2024-11-28 SevgiAkten/pycellga#79 @SevgiAkten
Dependency Analysis: Optimize and Clarify Requirements
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322

### Dependency Analysis for `requirements.txt`

Below is a detailed analysis of the dependencies listed in `requirements.txt`, including their purposes and relevance to the `pycellga` package.

---

#### **Essential Dependencies**
- **`numpy==1.26.2`**
  _Usage:_ Core dependency for numerical operations, matrix manipulations, and fundamental computations across all optimization problems.
  _Importance:_ **Critical**

- **`mpmath==1.3.0`**
  _Usage:_ Provides high-precision mathematical computations used in specific real-valued optimization problems like `Bohachevsky` and  `Bentcigar`.
  _Importance:_ **Critical**

- **`pymoo==0.5.0`**
  _Usage:_ Enables compatibility with the `pymoo` framework, allowing comparisons with other evolutionary algorithm implementations and access to a broader range of test problems.
  _Importance:_ **Critical**
_Implementation:_ Integrated following reviewer feedback during the review process ([Issue #71: Centralize Problem Configuration and Integrate pymoo Compatibility](https://github.com/SevgiAkten/pycellga/issues/71)), utilizing the `pymoo` `Problem` class to enhance functionality and enable compatibility with a wider range of test problems.

- **`geographiclib==2.0`** and **`geopy==2.4.1`**
  _Usage:_ Utilized specifically for the Traveling Salesman Problem (TSP) example to calculate distances between cities.
  _Importance:_ **Essential** for TSP-related functionalities.

- **`tsplib95==0.7.1`**
  _Usage:_ Utilized for handling TSP (Traveling Salesman Problem) data, specifically for loading benchmark datasets like `burma14`.
  _Importance:_ **Critical** for implementing TSP-related features and ensuring compatibility with standard problem datasets.
---

#### **Documentation and Testing Dependencies**
- **`Sphinx==7.4.7` and `sphinx-rtd-theme==2.0.0`**
  _Usage:_ Generates the project documentation in a clear, accessible format.
  _Importance:_ **Important** for maintaining high-quality documentation.

- **`pytest==8.3.2` and `pytest-xdist==3.6.1`**
  _Usage:_ Enables automated testing and parallel test execution, ensuring the reliability of `pycellga`.
  _Importance:_ **Important** for robust development and continuous integration.

---

#### **Candidates for Removal**
The following dependencies appear non-essential for the core functionality of `pycellga` and could potentially be removed to streamline the installation process:
- **`Markdown==3.6`** and **`Markdown-Editor==1.0.7`**
  _Usage:_ Used for rendering markdown files into HTML and editing markdown files.

- **`path-and-address==2.0.1`** and **`pluggy==1.5.0`**
  _Usage:_ Dependencies not explicitly used in the package’s core functionalities.

- **`grip==4.6.2`**
  _Usage:_ Used for rendering Markdown files as HTML locally

---
