### 2024-11-28 358074a
(#79) Remove non-essential optional dependencies
