+++ setup.cfg
@@ -23,14 +23,14 @@ python_requires = >=3.6
 install_requires =
     geographiclib==2.0
     geopy==2.4.1
-    grip==4.6.2
+    # grip==4.6.2 # Used for rendering Markdown files as HTML locally, not required for pycellga's core functionality
     iniconfig==2.0.0
-    Markdown==3.6
-    Markdown-Editor==1.0.7
+    # Markdown==3.6 # Used for rendering markdown files into HTML for documentation purposes
+    # Markdown-Editor==1.0.7 # Used for editing markdown files, not required for package functionality
     mpmath==1.3.0
     numpy==1.26.2
-    path-and-address==2.0.1
-    pluggy==1.5.0
+    # path-and-address==2.0.1 # Likely used for path handling, but not necessary for the core functionality of pycellga
+    # pluggy==1.5.0 # A dependency for pytest or other plugins, but not directly required for pycellga's core functionality
     plumbum==1.8.3
     ply==3.11
     pyparsing==3.1.1
