### 2024-11-28 SevgiAkten/pycellga#80 @SevgiAkten
Add Testing Guidelines for Improved Clarity
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
