### 2024-11-28 390ca98
(#80) Add Testing Guidelines to CONTRIBUTING and Documentation
