### 2024-12-05 SevgiAkten/pycellga#65 @jbussemaker
Hi @SevgiAkten, thanks for the many improvements over the past weeks. I would like to see the following further improvements:
- [ ] Manage releases in Github (you see the "Releases" header on the right pane when viewing the repository?)
- [ ] The documentation should still be improved:
  - [ ] There are only 4 very simple usage examples, you could include more and more complex examples
  - [ ] The examples are not consistent: in the first you use `result[1]` and `result[0]`, in the others you use `result.chromosome` and `result.fitness_value`
  - [ ] The "comprehensive guide" is just an API reference, please call it like that
- [ ] Everything is imported from `pycellga.optimizer`, even the test problems, this seems counterintuitive and makes it harder to understand how to use the code
- [ ] You must restructure the code as to place everything related to the problem definition together in the problem class!
  - [ ] It seems like this work was already started (in `abstract_problem.py`), however the documentation does not reflect this
  - [ ] The problem definition should include: gene (design variable) definitions (types, mins/maxs), the objective and constraint functions (`f` and `g`)
- [ ] You must use absolute module imports (see original comment)
- [ ] Thank you for changing to setup.cfg, but please improve the following:
  - [ ] Remove the redundant `requirements.txt`
  - [ ] Your dependency versions are way to strict, there's definitely not a need to fix all dependencies to a specific version (that would probably cause a lot of negative interactions if pycellga is used in an environment together with other libraries!)
  - [ ] Your dependency list contains packages probably not needed to run pycellga, such as six (six is needed for Python 2 compatibility, but you declare that pycellga can only be used by Python 3.6 and higher...), Sphinx
