### 2024-12-09 fefcd1a
(#69) Update main.yml to automate PyPI release and GitHub release
