### 2024-12-09 90f0016
(#70) Update PyPI badge for consistent styling
