+++ docs/pycellga.recombination.rst
@@ -6,7 +6,7 @@ The `pycellga.recombination` package includes various recombination (crossover) 
 
 
 **Arithmetic Crossover**
---------------
+-------------------------
 
 Performs arithmetic operations on parent genes to produce offspring. Often used in real-valued genetic algorithms to create intermediate values between parent genes.
 
@@ -16,7 +16,7 @@ Performs arithmetic operations on parent genes to produce offspring. Often used 
    :show-inheritance:
 
 **BLX-Alpha Crossover**
---------------
+------------------------
 
 Generates offspring by creating genes within a specified range around parent genes, controlled by the `alpha` parameter. Effective in real-valued optimization for maintaining diversity.
 
@@ -26,7 +26,7 @@ Generates offspring by creating genes within a specified range around parent gen
    :show-inheritance:
 
 **Byte One-Point Crossover**
---------------
+------------------------------
 
 A one-point crossover operator specifically designed for byte-based chromosomes, suitable for machine-coded genetic algorithms.
 
@@ -36,7 +36,7 @@ A one-point crossover operator specifically designed for byte-based chromosomes,
    :show-inheritance:
 
 **Byte Uniform Crossover**
---------------
+---------------------------
 
 Applies uniform crossover at the byte level, allowing fine control over gene mixing in machine-coded chromosomes.
 
@@ -46,7 +46,7 @@ Applies uniform crossover at the byte level, allowing fine control over gene mix
    :show-inheritance:
 
 **Flat Crossover**
---------------
+-------------------
 
 Creates offspring by generating random values within a range defined by the parent genes. Suitable for real-valued chromosomes.
 
@@ -56,7 +56,7 @@ Creates offspring by generating random values within a range defined by the pare
    :show-inheritance:
 
 **Linear Crossover**
---------------
+---------------------
 
 Generates offspring by linearly combining genes of the parents. This operator is useful in real-valued optimization for exploring intermediate values.
 
@@ -66,7 +66,7 @@ Generates offspring by linearly combining genes of the parents. This operator is
    :show-inheritance:
 
 **One-Point Crossover**
---------------
+------------------------
 
 A classic crossover technique where a single crossover point is chosen to swap segments between parents. Commonly used in binary-encoded genetic algorithms.
 
@@ -76,7 +76,7 @@ A classic crossover technique where a single crossover point is chosen to swap s
    :show-inheritance:
 
 **Partially Matched Crossover (PMX)**
---------------
+---------------------------------------
 
 A crossover method designed for permutation-based chromosomes, such as sequencing problems. Maintains gene order by partially matching segments between parents.
 
@@ -86,7 +86,7 @@ A crossover method designed for permutation-based chromosomes, such as sequencin
    :show-inheritance:
 
 **Two-Point Crossover**
---------------
+------------------------
 
 A crossover method with two crossover points, allowing for a higher level of gene exchange between parents. Widely used in binary-encoded algorithms.
 
@@ -96,7 +96,7 @@ A crossover method with two crossover points, allowing for a higher level of gen
    :show-inheritance:
 
 **Unfair Average Crossover**
---------------
+------------------------------
 
 Calculates the average of parent genes with a slight bias, leading to offspring that retain more traits of one parent. Used in real-valued genetic algorithms.
 
@@ -106,7 +106,7 @@ Calculates the average of parent genes with a slight bias, leading to offspring 
    :show-inheritance:
 
 **Uniform Crossover**
---------------
+----------------------
 
 Swaps genes randomly between parents, creating diverse offspring. Suitable for both binary and real-valued genetic representations.
 
