+++ _sources/pycellga.problems.single_objective.discrete.permutation.rst.txt
@@ -5,7 +5,7 @@ The `pycellga.problems.single_objective.discrete.permutation` package provides a
 
 
 **Traveling Salesman Problem (TSP)**
---------------
+----------------------------------------------
 
 The Traveling Salesman Problem (TSP) is a classic optimization problem where the goal is to find the shortest possible route that visits each city exactly once and returns to the origin city. TSP is widely used to evaluate the efficiency of algorithms that handle combinatorial and sequencing problems. This module provides a standard implementation of TSP for testing purposes.
 

[…]
