### 2024-12-10 8dac97b
(#68) Rename "Comprehensive Guide" to "API Reference" for accuracy
