### 2024-12-10 SevgiAkten/pycellga#81 @SevgiAkten
Refactor imports for improved modularity and code clarity
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
