+++ pycellga/example/example_cga.py
@@ -1,9 +1,14 @@
 import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-import optimizer
-from individual import GeneType
 from numpy import power as pw
+
+from optimizer import cga
+from individual import GeneType
+from recombination import ByteOnePointCrossover
+from mutation import ByteMutationRandom
+from selection import TournamentSelection
+
 
 class ExampleProblem:
     """
@@ -51,7 +56,7 @@ def run_cga_example():
     # Create an instance of the problem
     problem_instance = ExampleProblem()
 
-    result = optimizer.cga(
+    result = cga(
         n_cols=5,
         n_rows=5,
         n_gen=100,
@@ -60,9 +65,9 @@ def run_cga_example():
         p_crossover=0.9,
         p_mutation=0.2,
         problem=problem_instance,  # Pass the ExampleProblem instance
-        selection=optimizer.TournamentSelection,
-        recombination=optimizer.ByteOnePointCrossover,
-        mutation=optimizer.ByteMutationRandom,
+        selection=TournamentSelection,
+        recombination=ByteOnePointCrossover,
+        mutation=ByteMutationRandom,
         mins=[-32.768] * 5,  # Minimum values for each gene
         maxs=[32.768] * 5,    # Maximum values for each gene
         seed_par=100
