+++ pycellga/recombination/__init__.py
@@ -10,3 +10,18 @@ from . import recombination_operator
 from . import two_point_crossover
 from . import unfair_avarage_crossover
 from . import uniform_crossover
+
+# ------------------------------ recombination ----------------------------- #
+from recombination.one_point_crossover import OnePointCrossover
+from recombination.pmx_crossover import PMXCrossover
+from recombination.two_point_crossover import TwoPointCrossover
+from recombination.uniform_crossover import UniformCrossover
+from recombination.byte_uniform_crossover import ByteUniformCrossover
+from recombination.byte_one_point_crossover import ByteOnePointCrossover
+from recombination.flat_crossover import FlatCrossover
+from recombination.arithmetic_crossover import ArithmeticCrossover
+from recombination.blxalpha_crossover import BlxalphaCrossover
+from recombination.linear_crossover import LinearCrossover
+from recombination.unfair_avarage_crossover import UnfairAvarageCrossover
+from recombination.recombination_operator import RecombinationOperator
+# -------------------------------------------------------------------------- #
