+++ pycellga/mutation/__init__.py
@@ -7,3 +7,15 @@ from . import mutation_operator
 from . import shuffle_mutation
 from . import swap_mutation
 from . import two_opt_mutation
+
+# -------------------------------- mutation -------------------------------- #
+from mutation.bit_flip_mutation import BitFlipMutation
+from mutation.byte_mutation import ByteMutation
+from mutation.byte_mutation_random import ByteMutationRandom
+from mutation.insertion_mutation import InsertionMutation
+from mutation.shuffle_mutation import ShuffleMutation
+from mutation.swap_mutation import SwapMutation
+from mutation.two_opt_mutation import TwoOptMutation
+from mutation.float_uniform_mutation import FloatUniformMutation
+from mutation.mutation_operator import MutationOperator
+# -------------------------------------------------------------------------- #
