+++ pycellga/selection/__init__.py
@@ -1,3 +1,9 @@
 from . import roulette_wheel_selection
 from . import selection_operator
 from . import tournament_selection
+
+# ------------------------------ selection --------------------------------- #
+from selection.roulette_wheel_selection import RouletteWheelSelection
+from selection.tournament_selection import TournamentSelection
+from selection.selection_operator import SelectionOperator
+# -------------------------------------------------------------------------- #

[…]
