### 2024-12-16 3060c12
(#71) (#81) (#68) edit problem configuration, imports, and documentation
