+++ pycellga/tests/test_shuffle_mutation.py
@@ -1,6 +1,7 @@
 from mutation.shuffle_mutation import ShuffleMutation
 from problems.single_objective.discrete.permutation.tsp import Tsp
-from individual import Individual, GeneType
+from individual import Individual
+from common import GeneType
 import random
 
 def test_shuffle_mutation():
