+++ pycellga/tests/test_individual.py
@@ -1,7 +1,8 @@
 import pytest
 from numpy import random
 import random as rd
-from individual import Individual, GeneType 
+from individual import Individual
+from common import GeneType
 
 
 @pytest.fixture
