+++ _sources/pycellga.rst.txt
@@ -1,7 +1,7 @@
-pycellga: API Reference
-========================
+Components of pycellga
+=======================
 
-The `pycellga` package is a Python library for implementing and testing cellular genetic algorithms (CGAs). This guide provides an in-depth look at each module, offering descriptions and use cases to help users understand and utilize the library effectively.
+`pycellga` is a comprehensive Python library designed for implementing and experimenting with **cellular genetic algorithms (CGAs)**. CGAs are evolutionary algorithms characterized by structured populations and localized interactions, making them suitable for a wide range of optimization problems. This guide provides a detailed breakdown of each module within `pycellga`, highlighting their purpose, key features, and practical use cases.
 
 .. toctree::
    :maxdepth: 1
@@ -14,8 +14,8 @@ The `pycellga` package is a Python library for implementing and testing cellular
    pycellga.example
 
 
-**Core Modules**
-------------------
+**Core Modules API Refences**
+------------------------------
 
 Population Management
 ----------------------
@@ -26,7 +26,6 @@ Handles the initialization and management of the population in CGA. It includes 
    :members:
    :undoc-members:
    :show-inheritance:
-
 
 
 Individual Representation
@@ -40,7 +39,6 @@ Represents an individual in the population, encapsulating attributes like the ch
    :show-inheritance:
 
 
-
 Grid Structure
 --------------
 
@@ -50,7 +48,6 @@ Defines the grid structure for the cellular genetic algorithm. The grid layout r
    :members:
    :undoc-members:
    :show-inheritance:
-
 
 
 Byte Operators
@@ -66,9 +63,125 @@ Implements low-level byte-based operations that support machine-coded genetic al
 
 
 Optimizer
---------------
+===========
 
-The core optimization module that manages the genetic algorithm's execution. This includes processes like selection, mutation, recombination, and evaluation of individuals. The `optimizer` module serves as the main interface for running different CGA variants.
+The `pycellga.optimizer` module is the core component of the library, responsible for managing the execution of cellular genetic algorithms (CGAs). It integrates key evolutionary processes such as selection, mutation, recombination, and evaluation. The optimizer module also provides flexibility to implement and experiment with different CGA variants, each suited to specific optimization challenges.
+
+
+
+**Cellular Genetic Algorithm (cga)**
+-------------------
+
+The standard Cellular Genetic Algorithm (cga) operates with a fixed neighborhood structure and asynchronous updates. This approach provides robust performance for a wide range of problems, leveraging local interactions to explore the search space efficiently.
+
+.. image:: images/cga.png
+   :scale: 50%
+   :align: center
+   :alt: Standard CGA Example
+
+Figure 1: Illustration of the standard CGA process.
+
+
+**Synchronous Cellular Genetic Algorithm (sync_cga)**
+-------------------
+
+Synchronous CGA (sync_cga) updates the entire population simultaneously in each generation. This ensures consistent progress across the population but may lead to premature convergence in some cases.
+
+
+**Alpha Male Cellular Genetic Algorithm (alpha_cga)**
+-------------------
+
+This strategy developed to enhance the performance of Cellular Genetic Algorithms is the Alpha-Male Cellular Genetic Algorithm (alpha_cga). Alpha-Male Genetic Algorithm divides individuals in a population into social groups. A social group consists of females that select the same alpha male. In each social group, one individual is labeled as the alpha male, while the rest are productive females.
+
[…]
