+++ _sources/pycellga.example.rst.txt
@@ -1,66 +1,0 @@
-Example Implementations in pycellga
-===================================
-
-The `pycellga.example` package includes example implementations of various cellular genetic algorithms (CGAs) and their variants. These examples demonstrate how to set up and run different CGA methods using the `pycellga` framework, providing users with practical insights into the application of these algorithms.
-
-Available Example Modules
--------------------------
-
-### Example: Alpha-CGA
-**Module**: `pycellga.example.example_alpha_cga`
-
-Demonstrates the Alpha Cellular Genetic Algorithm (Alpha-CGA), an enhanced version of CGA that incorporates an "alpha male" mechanism for improved convergence and exploration.
-
-.. automodule:: pycellga.example.example_alpha_cga
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-### Example: Compact-CGA (CCGA)
-**Module**: `pycellga.example.example_ccga`
-
-Provides an example of the Compact Cellular Genetic Algorithm (CCGA), which utilizes a compact population structure to minimize memory usage and improve computational efficiency.
-
-.. automodule:: pycellga.example.example_ccga
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-### Example: Standard CGA
-**Module**: `pycellga.example.example_cga`
-
-Illustrates the implementation of a standard Cellular Genetic Algorithm (CGA), highlighting the core principles of CGA with basic settings.
-
-.. automodule:: pycellga.example.example_cga
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-### Example: Machine-Coded Compact CGA (MCC-CGA)
-**Module**: `pycellga.example.example_mcccga`
-
-Demonstrates the Machine-Coded Compact Cellular Genetic Algorithm (MCC-CGA), which uses byte-level encoding for efficient real-valued optimization in a compact population setting.
-
-.. automodule:: pycellga.example.example_mcccga
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-### Example: Synchronous CGA (Sync-CGA)
-**Module**: `pycellga.example.example_sync_cga`
-
-Illustrates the Synchronous Cellular Genetic Algorithm (Sync-CGA), where all individuals in the population are updated simultaneously, providing an alternative synchronization mechanism in CGA.
-
-.. automodule:: pycellga.example.example_sync_cga
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-## Full Module Overview
-
-For a comprehensive overview of the example implementations available in the `example` package, refer to the module documentation below:
-
-.. automodule:: pycellga.example
-   :members:
-   :undoc-members:
-   :show-inheritance:
