### 2024-12-16 af92c0c
Release version v0.3.4
