### 2024-12-16 SevgiAkten/pycellga#82 @SevgiAkten
Simplify and Clean Dependency List in setup.cfg
This is linked to the review feedback in Issue https://github.com/SevgiAkten/pycellga/issues/65 and part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7322
