### 2024-12-16 eb869fd
(#82) (#79) Cleaned dependencies and relaxed version constraints in setup.cfg.
