+++ setup.cfg
@@ -21,36 +21,19 @@ classifiers =
 packages = find:
 python_requires = >=3.6
 install_requires =
-    geographiclib==2.0
-    geopy==2.4.1
-    # grip==4.6.2 # Used for rendering Markdown files as HTML locally, not required for pycellga's core functionality
-    iniconfig==2.0.0
-    # Markdown==3.6 # Used for rendering markdown files into HTML for documentation purposes
-    # Markdown-Editor==1.0.7 # Used for editing markdown files, not required for package functionality
-    mpmath==1.3.0
-    numpy==1.26.2
-    # path-and-address==2.0.1 # Likely used for path handling, but not necessary for the core functionality of pycellga
-    # pluggy==1.5.0 # A dependency for pytest or other plugins, but not directly required for pycellga's core functionality
-    plumbum==1.8.3
-    ply==3.11
-    pyparsing==3.1.1
-    pytest==8.3.2
-    pytest-xdist==3.6.1
-    python-dateutil==2.8.2
-    pytz==2023.3.post1
-    six==1.16.0
-    Sphinx==7.4.7
-    sphinx-rtd-theme==2.0.0
-    sphinxcontrib-applehelp==2.0.0
-    sphinxcontrib-devhelp==2.0.0
-    sphinxcontrib-htmlhelp==2.1.0
-    sphinxcontrib-jquery==4.1
-    sphinxcontrib-jsmath==1.0.1
-    sphinxcontrib-qthelp==2.0.0
-    sphinxcontrib-serializinghtml==2.0.0
-    tsplib95==0.7.1
-    tzdata==2023.3
-    pymoo==0.5.0
+    geographiclib>=2.0,<3.0
+    geopy>=2.4.0,<3.0
+    iniconfig>=2.0.0,<3.0
+    mpmath>=1.3.0,<2.0
+    numpy>=1.26.0,<2.0
+    plumbum>=1.8.3,<2.0
+    ply>=3.11,<4.0
+    pyparsing>=3.1.1,<4.0
+    python-dateutil>=2.8.0,<3.0
+    pytz>=2023.3,<2024.0
+    tzdata>=2023.3,<2024.0
+    pymoo>=0.5.0,<1.0
+    tsplib95>=0.7.0,<1.0
 
 [options.packages.find]
 where = .
