+++ .github/workflows/documentation.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt 
+          pip install .
 
       - name: Build Docs
         run: |
