+++ .github/workflows/documentation.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .
+          pip install sphinx sphinx-rtd-theme
 
       - name: Build Docs
         run: |
