+++ .github/workflows/test.yml
@@ -19,10 +19,11 @@ jobs:
         with:
           python-version: '3.12'
 
-      - name: Install dependencies
+      - name: Install project dependencies
         run: |
           python -m pip install --upgrade pip
           pip install pytest
+          pip install .
 
       - name: Run tests
         run: pytest -v
