+++ .github/workflows/documentation.yml
@@ -22,7 +22,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install sphinx sphinx-rtd-theme
+          pip install Sphinx sphinx-rtd-theme \
+                sphinxcontrib-applehelp sphinxcontrib-htmlhelp \
+                sphinxcontrib-jquery sphinxcontrib-jsmath \
+                sphinxcontrib-qthelp sphinxcontrib-serializinghtml
 
       - name: Build Docs
         run: |
