+++ setup.cfg
@@ -35,6 +35,17 @@ install_requires =
     pymoo>=0.5.0,<1.0
     tsplib95>=0.7.0,<1.0
 
+[options.extras_require]
+docs =
+    sphinx>=7.4.0,<8.0
+    sphinx-rtd-theme>=2.0.0,<3.0
+    sphinxcontrib-applehelp>=2.0.0,<3.0
+    sphinxcontrib-htmlhelp>=2.1.0,<3.0
+    sphinxcontrib-jquery>=4.1,<5.0
+    sphinxcontrib-jsmath>=1.0.0,<2.0
+    sphinxcontrib-qthelp>=2.0.0,<3.0
+    sphinxcontrib-serializinghtml>=2.0.0,<3.0
+
 [options.packages.find]
 where = .
 
