+++ .github/workflows/documentation.yml
@@ -22,10 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install Sphinx sphinx-rtd-theme \
-                sphinxcontrib-applehelp sphinxcontrib-htmlhelp \
-                sphinxcontrib-jquery sphinxcontrib-jsmath \
-                sphinxcontrib-qthelp sphinxcontrib-serializinghtml
+          pip install .[docs]
 
       - name: Build Docs
         run: |
