+++ setup.cfg
@@ -36,6 +36,16 @@ install_requires =
     tsplib95>=0.7.0,<1.0
 
 [options.extras_require]
+dev =
+    pytest>=8.0,<9.0
+    pytest-xdist>=3.0,<4.0
+    grip==4.6.2
+    Markdown==3.6
+    Markdown-Editor==1.0.7
+    path-and-address==2.0.1
+    pluggy==1.5.0
+    six==1.16.0
+
 docs =
     sphinx>=7.4.0,<8.0
     sphinx-rtd-theme>=2.0.0,<3.0
