+++ .github/workflows/documentation.yml
@@ -22,6 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install .[dev]
           pip install .[docs]
 
       - name: Build Docs
