+++ pycellga/problems/single_objective/continuous/rosenbrock.py
@@ -1,7 +1,8 @@
-from problems.abstract_problem import AbstractProblem
-from mpmath import power as pw
-from common import GeneType
 from typing import List
+from mpmath import power as pw
+
+from pycellga.problems.abstract_problem import AbstractProblem
+from pycellga.common import GeneType
 
 
 class Rosenbrock(AbstractProblem):
