+++ pycellga/problems/single_objective/continuous/levy.py
@@ -1,8 +1,10 @@
-from problems.abstract_problem import AbstractProblem
+
 import math
 from mpmath import power as pw
 from typing import List
-from common import GeneType
+
+from pycellga.problems.abstract_problem import AbstractProblem
+from pycellga.common import GeneType
 
 
 class Levy(AbstractProblem):
