+++ pycellga/recombination/two_point_crossover.py
@@ -1,8 +1,9 @@
 import numpy as np
-from individual import Individual
-from problems.abstract_problem import AbstractProblem
 from typing import List
-from recombination.recombination_operator import RecombinationOperator
+
+from pycellga.individual import Individual
+from pycellga.problems.abstract_problem import AbstractProblem
+from pycellga.recombination.recombination_operator import RecombinationOperator
 
 class TwoPointCrossover(RecombinationOperator):
     """
