### 2024-12-17 #7322 @jbussemaker
Hi @SevgiAkten, thank you for your latest improvements. The code quality and documentation have improved greatly. Especially for the documentation I want to compliment you on the addition of many figures to greatly help to explain the code and theory behind it. Also the package management (setup.cfg, releases, PyPI, etc.) has improved greatly.

@JBorrow I do not have any further comments either, and I recommend the paper for publication as well.
