### 2024-12-21 #7322 @SevgiAkten
Dear @JBorrow,
I have completed the required archiving process for the software on Zenodo. Below are the details:

DOI: [10.5281/zenodo.14539107](https://doi.org/10.5281/zenodo.14539107)

Zenodo Link: https://zenodo.org/record/14539107

Version Number: v0.4.0
