### 2025-01-02 #7322 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
