### 2025-01-02 #7322 @crvernon
:wave: @SevgiAkten - thank you for the submission.  I just need you to address the following before I can move to accept this for publication:

- [ ] LINE 28:  "...by (Satman, 2013)." should be referenced as "...by Satman (2013)."
- [ ] LINE 62:  "...by (Satman, 2013)." should be referenced as "...by Satman (2013)."
- [ ] LINE 65:  "...by (Satman, 2013)." should be referenced as "...by Satman (2013)."
- [ ] LINE 73:  "...from (Satman & Akadal, 2019)." should be referenced as "...from Satman & Akadal (2019)."
- [ ] LINES 73-74:  "...by(Satman & Akadal, 2020)." should be referenced as "...by Satman & Akadal (2020)."
- [ ] LINE 126-127:  "...by (Satman, 2013)," should be referenced as "...by Satman (2013),"
- [ ] LINE 145:  "cga" should be capitalized, or if you are meaning to call out the specific method (which I think you are) place "cga" in backticks to give it a code formatting or bold it as you do later in the paper.  Also, there is a missing "the" should be added in as "...explain the CGA method..."  You should make the formatting of `cga` as a module consistent throughout the "Installation and basic usage" section for whichever formatting you prefer.
- [ ] There is no need to keep using the full breakout of the acronym for CGA after you have initially defined it.  Feel free to reduce subsequent uses of "Cellular Genetic Algorithm (CGA)" to simply use "CGA"

Let me know once you have made these changes and I'll accept this for publication.  Thanks!
