+++ paper/paper.md
@@ -29,7 +29,7 @@ While CGAs themselves are not a novel contribution of this work, `pycellga` sign
 
 The package is designed to be user-friendly, with a straightforward installation process and comprehensive documentation. Researchers and practitioners in fields such as operations research, artificial intelligence, and machine learning can leverage `pycellga` to tackle complex optimization challenges effectively. By integrating the principles of cellular automata with genetic algorithms, `pycellga` represents a significant advancement in the field of evolutionary computation, offering increased flexibility and adaptability compared to traditional methods.
 
-Additionally, `pycellga` includes machine-coded operators with byte implementations, developed by Satman (2013) [@satman2013machine]. It features Alpha-male CGA, Machine-Coded Compact CGA, and Improved CGA with Machine-Coded Operators for real-valued optimization problems [@karakaya2024improved].
+Additionally, `pycellga` includes machine-coded operators with byte implementations, developed by @satman2013machine. It features Alpha-male CGA, Machine-Coded Compact CGA, and Improved CGA with Machine-Coded Operators for real-valued optimization problems [@karakaya2024improved].
 
 
 # Introduction
@@ -42,9 +42,9 @@ Despite their effectiveness, traditional GAs face challenges, particularly in ma
 
 By providing a comprehensive toolkit for CGAs, `pycellga` aims to advance the field of evolutionary computation and equip researchers with the tools needed to solve increasingly complex optimization problems effectively. The integration of cellular automata with genetic algorithms in `pycellga` represents a significant advancement, offering greater flexibility and adaptability compared to traditional methods [@michalewicz1996genetic; @eiben2003introduction; @karakaya2024improved]. 
 
-The `pycellga` package includes machine-coded operators with byte-level implementations, developed by Satman (2013) [@satman2013machine]. In the context of genetic algorithms, "machine-coded" refers to a specialized encoding technique optimized for real-parameter optimization. This approach differs from standard coding practices by emphasizing efficient data processing through byte-level manipulation. Originally introduced by Satman (2013) [@satman2013machine], this technique is particularly advantageous for real-valued optimization tasks, as it allows direct manipulation of byte-representations to enhance computational performance. Encoding and decoding of numerical values conform to the IEEE 754 standard for floating-point arithmetic, further improving precision and effectiveness in optimizing continuous functions. By using machine-coded operators, `pycellga` leverages this efficiency to handle complex optimization challenges more effectively.
+The `pycellga` package includes machine-coded operators with byte-level implementations, developed by @satman2013machine. In the context of genetic algorithms, "machine-coded" refers to a specialized encoding technique optimized for real-parameter optimization. This approach differs from standard coding practices by emphasizing efficient data processing through byte-level manipulation. Originally introduced by @satman2013machine, this technique is particularly advantageous for real-valued optimization tasks, as it allows direct manipulation of byte-representations to enhance computational performance. Encoding and decoding of numerical values conform to the IEEE 754 standard for floating-point arithmetic, further improving precision and effectiveness in optimizing continuous functions. By using machine-coded operators, `pycellga` leverages this efficiency to handle complex optimization challenges more effectively.
 
[…]
